+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 1.3 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Add support for context portlets.
+  [buchi]
 
 
 1.2 (2016-05-24)